Task 15 - DOM
Specifically, proceed to implement the DOMBuilder class, with, of
course, support for attributes. Inasmuch as this class extends our
DefaultHandler, it is designed to build a DOM out of SAX events. As
suggested by DOMBuilder.java, that DOM will be represented by a
reference to a Document object whose descendents are instances of the
Attr, Element, andTextclasses.
Alloftheseobjects,mindyou,arealsooftypeNode.
To be clear, in addition to modifying DOMBuilder.java, be sure to alter
Attr.java and Element.java as you see fit.5 You are welcome to declare
and define other classes; be sure they belong to the
cscie259.project1.mf package.